Improve random, export common utility to testbase.rs
