+++ src/bvh.rs
@@ -346,36 +346,19 @@ impl BVH {
 
 #[cfg(test)]
 pub mod tests {
-    use aabb::{AABB, Bounded};
-    use bvh::BVH;
-    use nalgebra::{Point3, Vector3};
     use std::collections::HashSet;
+
+    use nalgebra::{Point3, Vector3};
+
+    use bvh::BVH;
     use ray::Ray;
-
-    /// Define some Bounded structure.
-    pub struct XBox {
-        x: i32,
-    }
-
-    /// `XBox`'s `AABB`s are unit `AABB`s centered on the given x-position.
-    impl Bounded for XBox {
-        fn aabb(&self) -> AABB {
-            let min = Point3::new(self.x as f32 - 0.5, -0.5, -0.5);
-            let max = Point3::new(self.x as f32 + 0.5, 0.5, 0.5);
-            AABB::with_bounds(min, max)
-        }
-    }
+    use testbase::{generate_aligned_boxes, UnitBox, create_n_cubes, create_ray};
 
     /// Creates a `BVH` for a fixed scene structure.
-    pub fn build_some_bvh() -> (Vec<XBox>, BVH) {
-        // Create 21 boxes along the x-axis
-        let mut shapes = Vec::new();
-        for x in -10..11 {
-            shapes.push(XBox { x: x });
-        }
-
-        let bvh = BVH::build(&shapes);
-        (shapes, bvh)
+    pub fn build_some_bvh() -> (Vec<UnitBox>, BVH) {
+        let boxes = generate_aligned_boxes();
+        let bvh = BVH::build(&boxes);
+        (boxes, bvh)
     }
 
     #[test]
@@ -384,136 +367,61 @@ pub mod tests {
         build_some_bvh();
     }
 
+    fn traverse_and_verify(ray_origin: Point3<f32>,
+                           ray_direction: Vector3<f32>,
+                           all_shapes: &Vec<UnitBox>,
+                           bvh: &BVH,
+                           expected_shapes: &HashSet<i32>) {
+        let ray = Ray::new(ray_origin, ray_direction);
+        let hit_shapes = bvh.traverse_recursive(&ray, all_shapes);
+
+        assert_eq!(expected_shapes.len(), hit_shapes.len());
+        for shape in hit_shapes {
+            assert!(expected_shapes.contains(&shape.id));
+        }
+    }
+
     #[test]
     /// Runs some primitive tests for intersections of a ray with a fixed scene given as a BVH.
     fn test_traverse_recursive_bvh() {
-        let (shapes, bvh) = build_some_bvh();
+        let (all_shapes, bvh) = build_some_bvh();
 
-        // Define a ray which traverses the x-axis from afar
-        let position_1 = Point3::new(-1000.0, 0.0, 0.0);
-        let direction_1 = Vector3::new(1.0, 0.0, 0.0);
-        let ray_1 = Ray::new(position_1, direction_1);
+        {
+            // Define a ray which traverses the x-axis from afar.
+            let origin = Point3::new(-1000.0, 0.0, 0.0);
+            let direction = Vector3::new(1.0, 0.0, 0.0);
+            let mut expected_shapes = HashSet::new();
 
-        // It shuold hit all shapes
-        let hit_shapes_1 = bvh.traverse_recursive(&ray_1, &shapes);
-        assert!(hit_shapes_1.len() == 21);
-        let mut xs_1 = HashSet::new();
-        for shape in &hit_shapes_1 {
-            xs_1.insert(shape.x);
-        }
-        for x in -10..11 {
-            assert!(xs_1.contains(&x));
-        }
-
-        // Define a ray which traverses the y-axis from afar
-        let position_2 = Point3::new(0.0, -1000.0, 0.0);
-        let direction_2 = Vector3::new(0.0, 1.0, 0.0);
-        let ray_2 = Ray::new(position_2, direction_2);
-
-        // It should hit only one box
-        let hit_shapes_2 = bvh.traverse_recursive(&ray_2, &shapes);
-        assert!(hit_shapes_2.len() == 1);
-        assert!(hit_shapes_2[0].x == 0);
-
-        // Define a ray which intersects the x-axis diagonally
-        let position_3 = Point3::new(6.0, 0.5, 0.0);
-        let direction_3 = Vector3::new(-2.0, -1.0, 0.0);
-        let ray_3 = Ray::new(position_3, direction_3);
-
-        // It should hit exactly three boxes
-        let hit_shapes_3 = bvh.traverse_recursive(&ray_3, &shapes);
-        assert!(hit_shapes_3.len() == 3);
-        let mut xs_3 = HashSet::new();
-        for shape in &hit_shapes_3 {
-            xs_3.insert(shape.x);
-        }
-        assert!(xs_3.contains(&6));
-        assert!(xs_3.contains(&5));
-        assert!(xs_3.contains(&4));
-    }
-
-    /// A triangle struct. Instance of a more complex `Bounded` primitive.
-    pub struct Triangle {
-        pub a: Point3<f32>,
-        pub b: Point3<f32>,
-        pub c: Point3<f32>,
-        aabb: AABB,
-    }
-
-    impl Triangle {
-        fn new(a: Point3<f32>, b: Point3<f32>, c: Point3<f32>) -> Triangle {
-            Triangle {
-                a: a,
-                b: b,
-                c: c,
-                aabb: AABB::empty().grow(&a).grow(&b).grow(&c),
+            // It should hit everything.
+            for id in -10..11 {
+                expected_shapes.insert(id);
             }
+            traverse_and_verify(origin, direction, &all_shapes, &bvh, &expected_shapes);
         }
-    }
 
-    impl Bounded for Triangle {
-        fn aabb(&self) -> AABB {
-            self.aabb
+        {
+            // Define a ray which traverses the y-axis from afar.
+            let origin = Point3::new(0.0, -1000.0, 0.0);
+            let direction = Vector3::new(0.0, 1.0, 0.0);
+
+            // It should hit only one box.
+            let mut expected_shapes = HashSet::new();
+            expected_shapes.insert(0);
+            traverse_and_verify(origin, direction, &all_shapes, &bvh, &expected_shapes);
         }
-    }
 
-    /// Creates a unit size cube centered at `pos` and pushes the triangles to `shapes`
-    fn push_cube(pos: Point3<f32>, shapes: &mut Vec<Triangle>) {
-        let top_front_right = pos + Vector3::new(0.5, 0.5, -0.5);
-        let top_back_right = pos + Vector3::new(0.5, 0.5, 0.5);
-        let top_back_left = pos + Vector3::new(-0.5, 0.5, 0.5);
-        let top_front_left = pos + Vector3::new(-0.5, 0.5, -0.5);
-        let bottom_front_right = pos + Vector3::new(0.5, -0.5, -0.5);
-        let bottom_back_right = pos + Vector3::new(0.5, -0.5, 0.5);
-        let bottom_back_left = pos + Vector3::new(-0.5, -0.5, 0.5);
-        let bottom_front_left = pos + Vector3::new(-0.5, -0.5, -0.5);
+        {
+            // Define a ray which intersects the x-axis diagonally.
+            let origin = Point3::new(6.0, 0.5, 0.0);
+            let direction = Vector3::new(-2.0, -1.0, 0.0);
 
-        shapes.push(Triangle::new(top_back_right, top_front_right, top_front_left));
-        shapes.push(Triangle::new(top_front_left, top_back_left, top_back_right));
-        shapes.push(Triangle::new(bottom_front_left, bottom_front_right, bottom_back_right));
-        shapes.push(Triangle::new(bottom_back_right, bottom_back_left, bottom_front_left));
-        shapes.push(Triangle::new(top_back_left, top_front_left, bottom_front_left));
-        shapes.push(Triangle::new(bottom_front_left, bottom_back_left, top_back_left));
-        shapes.push(Triangle::new(bottom_front_right, top_front_right, top_back_right));
-        shapes.push(Triangle::new(top_back_right, bottom_back_right, bottom_front_right));
-        shapes.push(Triangle::new(top_front_left, top_front_right, bottom_front_right));
-        shapes.push(Triangle::new(bottom_front_right, bottom_front_left, top_front_left));
-        shapes.push(Triangle::new(bottom_back_right, top_back_right, top_back_left));
-        shapes.push(Triangle::new(top_back_left, bottom_back_left, bottom_back_right));
-    }
-
-    /// Implementation of splitmix64.
-    /// For reference see: http://xoroshiro.di.unimi.it/splitmix64.c
-    fn splitmix64(x: &mut u64) -> u64 {
-        *x = x.wrapping_add(0x9E3779B97F4A7C15u64);
-        let mut z = *x;
-        z = (z ^ (z >> 30)).wrapping_mul(0xBF58476D1CE4E5B9u64);
-        z = (z ^ (z >> 27)).wrapping_mul(0x94D049BB133111EBu64);
-        z ^ (z >> 31)
-    }
-
-    fn next_point3(seed: &mut u64) -> Point3<f32> {
-        let u = splitmix64(seed);
-        let a = (u >> 48 & 0xFFFFFFFF) as i32 - 0xFFFF;
-        let b = (u >> 48 & 0xFFFFFFFF) as i32 - 0xFFFF;
-        let c = a ^ b.rotate_left(6);
-        Point3::new(a as f32, b as f32, c as f32)
-    }
-
-    pub fn create_n_cubes(n: u64) -> Vec<Triangle> {
-        let mut vec = Vec::new();
-        let mut seed = 0;
-
-        for _ in 0..n {
-            push_cube(next_point3(&mut seed), &mut vec);
+            // It should hit exactly three boxes.
+            let mut expected_shapes = HashSet::new();
+            expected_shapes.insert(4);
+            expected_shapes.insert(5);
+            expected_shapes.insert(6);
+            traverse_and_verify(origin, direction, &all_shapes, &bvh, &expected_shapes);
         }
-        vec
-    }
-
-    pub fn create_ray(seed: &mut u64) -> Ray {
-        let origin = next_point3(seed);
-        let direction = next_point3(seed).to_vector();
-        Ray::new(origin, direction)
     }
 
     #[bench]
@@ -523,41 +431,6 @@ pub mod tests {
 
         b.iter(|| {
             BVH::build(&triangles);
-        });
-    }
-
-    #[bench]
-    /// Benchmark intersecting 120,000 triangles directly.
-    fn bench_intersect_120k_triangles_list(b: &mut ::test::Bencher) {
-        let triangles = create_n_cubes(10_000);
-        let mut seed = 0;
-
-        b.iter(|| {
-            let ray = create_ray(&mut seed);
-
-            // Iterate over the list of triangles
-            for triangle in &triangles {
-                ray.intersects_triangle(&triangle.a, &triangle.b, &triangle.c);
-            }
-        });
-    }
-
-    #[bench]
-    /// Benchmark intersecting 120,000 triangles with preceeding `AABB` tests.
-    fn bench_intersect_120k_triangles_list_aabb(b: &mut ::test::Bencher) {
-        let triangles = create_n_cubes(10_000);
-        let mut seed = 0;
-
-        b.iter(|| {
-            let ray = create_ray(&mut seed);
-
-            // Iterate over the list of triangles
-            for triangle in &triangles {
-                // First test whether the ray intersects the AABB of the triangle
-                if ray.intersects_aabb(&triangle.aabb()) {
-                    ray.intersects_triangle(&triangle.a, &triangle.b, &triangle.c);
-                }
-            }
         });
     }
